4 Scrolling the Scene
two changes made. First the method is checking to see whether the
playerNode has moved at least 180 points up the scene. If the
playerNode has moved this high, then the backgroundNode is moved down
the scene at 1/8th the speed of the player, and then the foreground is
moved at exactly the same rate as the player.